requirements.txt
opencv-python
pysrt
pillow
piexif
scikit-image